Increase number of supported edge to >trillions (#84)
* Increase number of supported edge to >trillions

Backwards-compatible file format change. Previously the CSR arrays
were stored with 32-bit indices, which limited the number of edges
that a GRG could have. They now use 64-bit indices.

I also put some hard constraints around node creation, and added
automated tests for these edge cases.

+++ test/unit/test_csr_storage.cpp
@@ -181,7 +181,7 @@ TEST(CSRStorageImm, Writing) {
     uint64_t valueBytesRd = 0;
     infile->read((char*)&valueBytesRd, sizeof(uint64_t));
     EagerFileVector<uint8_t> edges(infile, sizeof(uint64_t), valueBytesRd);
-    EagerFileVector<uint32_t> nodes(infile, sizeof(uint64_t)+valueBytesRd, numNodes+1);
+    EagerFileVector<uint64_t> nodes(infile, sizeof(uint64_t)+valueBytesRd, numNodes+1);
     CSRStorageImm<EagerFileVector, uint32_t, true, true> compressedRead(
         std::move(nodes), std::move(edges), 0);
     for (size_t i = 0; i < numNodes; i++) {
